update all int and floats to 64

+++ zego/comments.go
@@ -9,23 +9,23 @@ type CommentArray struct {
 }
 
 type Comments struct {
-	Id          int           `json:"id"`
+	Id          int64         `json:"id"`
 	Type        string        `json:"type"`
 	Body        string        `json:"body"`
 	Public      bool          `json:"public"`
 	CreatedAt   string        `json:"created_at"`
-	AuthorId    int           `json:"author_id"`
+	AuthorId    int64         `json:"author_id"`
 	Attachments []Attachments `json:"attachments"`
 	Metadata    interface{}   `json:"metadata"`
 	Via         interface{}   `json:"via"`
 }
 
 type Attachments struct {
-	Id          int           `json:"id"`
+	Id          int64         `json:"id"`
 	Name        string        `json:"name"`
 	ContentURL  string        `json:"content_url"`
 	ContentType string        `json:"content_type"`
-	Size        int           `json:"size"`
+	Size        int64         `json:"size"`
 	Thumbnails  []interface{} `json:"thumbnails"`
 }
